dorilarimuz/requirements.txt
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
requests==2.27.1
path==16.4.0
Pillow==9.1.1
qrcode==7.3.1
fpdf2==2.5.1